fastapi/requirements.txt
fastapi==0.103.0
pillow==10.0.0
torch==2.0.1+rocm5.4.2
torchvision==0.15.2+rocm5.4.2
python-multipart==0.0.6
aiofiles==23.2.1
albumentations==1.3.1